+++ lib/additionals/patches/issue_patch.rb
@@ -145,9 +145,13 @@ module Additionals
       end
 
       def validate_change_on_closed
-        return true if !closed? || new_record? || !Additionals.setting?(:issue_freezed_with_close)
+        return true if !closed? ||
+                       new_record? ||
+                       !Additionals.setting?(:issue_freezed_with_close) ||
+                       !status_was.is_closed ||
+                       User.current.allowed_to?(:edit_closed_issues, project)
 
-        errors.add :base, :issue_changes_not_allowed unless User.current.allowed_to?(:edit_closed_issues, project)
+        errors.add :base, :issue_changes_not_allowed
       end
 
       def validate_open_sub_issues